@@ -1,19 +1,15 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <LangVersion>Latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MassTransit.Extensions.DependencyInjection" Version="5.5.6" />
-    <PackageReference Include="MassTransit.Extensions.Logging" Version="5.5.6" />
-    <PackageReference Include="MassTransit.RabbitMQ" Version="5.5.6" />
-    <PackageReference Include="MassTransit.Azure.ServiceBus.Core" Version="5.5.6" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="NSwag.AspNetCore" Version="13.0.6" />
+    <PackageReference Include="MassTransit.Extensions.DependencyInjection" Version="6.0.1" />
+    <PackageReference Include="MassTransit.RabbitMQ" Version="6.0.1" />
+    <PackageReference Include="MassTransit.Azure.ServiceBus.Core" Version="6.0.1" />
+    <PackageReference Include="NSwag.AspNetCore" Version="13.2.2" />
   </ItemGroup>
 
   <ItemGroup>
