@@ -1,5 +1,5 @@
 ﻿using MassTransit;
-using MassTransit.Logging.Tracing;
+using MassTransit.Context;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using System.Threading;
@@ -15,9 +15,6 @@ namespace SampleBatch.Service
         public MassTransitConsoleHostedService(IBusControl bus, ILoggerFactory loggerFactory)
         {
             _bus = bus;
-
-            if (loggerFactory != null && MassTransit.Logging.Logger.Current.GetType() == typeof(TraceLogger))
-                MassTransit.ExtensionsLoggingIntegration.ExtensionsLogger.Use(loggerFactory);
         }
 
         public async Task StartAsync(CancellationToken cancellationToken)
